Blasr: move MappingBuffers and ReadAlignment classes.
