+++ Blasr.cpp
@@ -1,3 +1,40 @@
+// Copyright (c) 2014-2015, Pacific Biosciences of California, Inc.
+//
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted (subject to the limitations in the
+// disclaimer below) provided that the following conditions are met:
+//
+//  * Redistributions of source code must retain the above copyright
+//    notice, this list of conditions and the following disclaimer.
+//
+//  * Redistributions in binary form must reproduce the above
+//    copyright notice, this list of conditions and the following
+//    disclaimer in the documentation and/or other materials provided
+//    with the distribution.
+//
+//  * Neither the name of Pacific Biosciences nor the names of its
+//    contributors may be used to endorse or promote products derived
+//    from this software without specific prior written permission.
+//
+// NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE
+// GRANTED BY THIS LICENSE. THIS SOFTWARE IS PROVIDED BY PACIFIC
+// BIOSCIENCES AND ITS CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
+// WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+// OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+// DISCLAIMED. IN NO EVENT SHALL PACIFIC BIOSCIENCES OR ITS
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF
+// USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+// ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
+// OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+// SUCH DAMAGE.
+
+// Author: Mark Chaisson
+
 #ifdef __linux__
 #  include <mcheck.h>
 #endif
@@ -15,6 +52,8 @@
 #include "MappingIPC.h"
 #include "MappingSemaphores.h"
 #include "RegisterBlasrOptions.h"
+#include "MappingBuffers.hpp"
+#include "ReadAlignments.hpp"
 
 #include "CCSSequence.hpp"
 #include "SMRTSequence.hpp"
@@ -127,153 +166,6 @@ ReaderAgglomerate *reader = NULL;
 
 typedef MappingData<T_SuffixArray, T_GenomeSequence, T_Tuple> MappingIPC;
 
-class ClusterInformation {
-public:
-  int maxClusterSize;
-  float meanAnchorBasesPerRead;
-  float sdAnchorBasesPerRead;
-  int score;
-  float pctSimilarity;
-  int readLength;
-  float nStdDev ;
-  int numSignificant;
-  int numForward, numReverse;
-};
-
-
-class ReadAlignments {
-public:
-  /*
-    This class stores the alignments from a read.  A read may be
-    aligned in several different modes:
-    1. Fullread    - Treat the read as a unit from start to end
-    2. Subread     - Align each subread independently
-    3. CCSDeNovo   - Only align the CCS sequence from a read
-    4. CCSAllPass  - Align the de novo ccs sequences and then the
-                     subreads to where the denovo ccs aligned.
-    5. CCSFullPass - Same as allpass, except using only complete
-                     subreads.
-    6. ZmwSubreads - Align subreads of each zmw to where the longest 
-                     subread of the zmw aligned to.
-   
-    The alignments are a raggad array of n sequences; n is 1 for cases 
-    1 and 3, the number of subreads for cases 2 and 4, and the number
-    of full length passes for case 5.
-
-    A ReadAligments class must only have alignments for a single type
-    of read in it.
-
-  */
-
-  vector<vector<T_AlignmentCandidate*> > subreadAlignments;
-  vector<SMRTSequence> subreads;
-  AlignMode alignMode;
-  SMRTSequence read;
-  int GetNAlignedSeq() {
-    return subreadAlignments.size();
-  }
-
-  bool AllSubreadsHaveAlignments() {
-    int i, nAlignedSeq;
-    nAlignedSeq = subreadAlignments.size();
-    for (i = 0; i < nAlignedSeq; i++) {
-      if (subreadAlignments[i].size() == 0) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  void Clear() {
-    int i;
-    int nAlignedSeq;
-    for (i = 0, nAlignedSeq = subreadAlignments.size(); i < nAlignedSeq; i++) {
-      int nAlignments;
-      int a;
-      for (a = 0, nAlignments = subreadAlignments[i].size(); a < nAlignments; a++) {
-        delete subreadAlignments[i][a];
-      }
-      subreadAlignments[i].clear();
-    }
-
-    for (i = 0, nAlignedSeq = subreads.size(); i< nAlignedSeq; i++) {
-      subreads[i].Free();
-    }
-    subreadAlignments.clear();
-    read.Free();
-  }
-
-  void Resize(int nSeq) {
-      subreadAlignments.resize(nSeq);
-      subreads.resize(nSeq);
-  }
-
-  void CheckSeqIndex(int seqIndex) {
-    if ( seqIndex < 0 or seqIndex >= int(subreads.size()) ) {
-      cout << "ERROR, adding a sequence to an unallocated position." 
-           << endl;
-      assert(0);
-    }
-  }
-
-  void SetSequence(int seqIndex, SMRTSequence &seq) {
-    CheckSeqIndex(seqIndex);
-    subreads[seqIndex] = seq;
-  }
-
-  void AddAlignmentForSeq(int seqIndex, T_AlignmentCandidate *alignmentPtr) {
-    CheckSeqIndex(seqIndex);
-    subreadAlignments[seqIndex].push_back(alignmentPtr);
-  }
-
-  void AddAlignmentsForSeq(int seqIndex, vector<T_AlignmentCandidate*> &seqAlignmentPtrs) {
-    CheckSeqIndex(seqIndex);
-    subreadAlignments[seqIndex].insert(subreadAlignments[seqIndex].end(), seqAlignmentPtrs.begin(), seqAlignmentPtrs.end());
-  }
-
-  // Copy all T_AlignmentCandidate objects (to which subreadAlignment[seqIndex]
-  // is pointing) to newly created objects, and then return pointers to the new
-  // objects.
-  vector<T_AlignmentCandidate*> CopySubreadAlignments(int seqIndex) {
-    vector<T_AlignmentCandidate*> ret;
-    for (int i=0; i<int(subreadAlignments[seqIndex].size()); i++) {
-      T_AlignmentCandidate * q = new T_AlignmentCandidate();
-      *q = *(subreadAlignments[seqIndex][i]);
-      ret.push_back(q);
-    }
-    return ret;
-  }
-
-  void Print(ostream &out=cout) { 
-    out << "A ReadAlignments object with " 
-        << subreadAlignments.size()
-        << " groups of subread alignments." << endl;
-        for (int i = 0; i < int(subreadAlignments.size()); i++) {
-            out << "  subreadAlignment group [" << i << "/" 
-                << subreadAlignments.size() << "] has "
-                << subreadAlignments[i].size() << " alignments." << endl;
-            for(int j = 0; j < int(subreadAlignments[i].size()); j++) {
-                out << "    [" << i << "][" << j << "/" 
-                    << subreadAlignments[i].size() << "]" << endl;
-                subreadAlignments[i][j]->Print(out);
-            } 
-        }
-        /* subreads may have been freed or not initialized when 
-             * being printed. 
-        for (int i = 0; i < subreads.size(); i++) {
-            out << "  subread [" << i << "/" << subreads.size()
-                << "]: ";
-            subreads[i].Print(out);
-        } */
-        out << "  read: ";
-        read.Print(out);
-        out << endl << endl;
-    }
-
-  ~ReadAlignments() {
-      read.Free();
-  }
-};
 
 string GetMajorVersion() {
   return "2.0.0";
@@ -289,68 +181,6 @@ const string GetVersion(void) {
   return version;
 }
 
-
-//
-// Define a list of buffers that are meant to grow to high-water
-// marks, and not shrink down past that.   The memory is reused rather
-// than having multiple calls to new.
-//
-class MappingBuffers {
-public:
-  vector<int> hpInsScoreMat, insScoreMat;
-  vector<int> kbandScoreMat;
-  vector<Arrow> hpInsPathMat, insPathMat;
-  vector<Arrow> kbandPathMat;
-  vector<int>   scoreMat;
-  vector<Arrow> pathMat;
-  vector<int>  affineScoreMat;
-  vector<Arrow> affinePathMat;
-  vector<ChainedMatchPos> matchPosList;
-  vector<ChainedMatchPos> rcMatchPosList;
-  vector<BasicEndpoint<ChainedMatchPos> > globalChainEndpointBuffer;
-  vector<Fragment> sdpFragmentSet, sdpPrefixFragmentSet, sdpSuffixFragmentSet;
-  TupleList<PositionDNATuple> sdpCachedTargetTupleList;
-  TupleList<PositionDNATuple> sdpCachedTargetPrefixTupleList;
-  TupleList<PositionDNATuple> sdpCachedTargetSuffixTupleList;
-  std::vector<int> sdpCachedMaxFragmentChain;
-  vector<double> probMat;
-  vector<double> optPathProbMat;
-  vector<float>  lnSubPValueMat;
-  vector<float>  lnInsPValueMat;
-  vector<float>  lnDelPValueMat;
-  vector<float>  lnMatchPValueMat;
-  vector<int>    clusterNumBases;
-  ClusterList    clusterList;
-  ClusterList    revStrandClusterList;
-
-  void Reset() {
-    vector<int>().swap(hpInsScoreMat);
-    vector<int>().swap(insScoreMat);
-    vector<int>().swap(kbandScoreMat);
-    vector<Arrow>().swap(hpInsPathMat);
-    vector<Arrow>().swap(insPathMat);
-    vector<Arrow>().swap(kbandPathMat);
-    vector<int>().swap(scoreMat);
-    vector<Arrow>().swap(pathMat);
-    vector<ChainedMatchPos>().swap(matchPosList);
-    vector<ChainedMatchPos>().swap(rcMatchPosList);
-    vector<BasicEndpoint<ChainedMatchPos> >().swap(globalChainEndpointBuffer);
-    vector<Fragment>().swap(sdpFragmentSet);
-    vector<Fragment>().swap(sdpPrefixFragmentSet);
-    vector<Fragment>().swap(sdpSuffixFragmentSet);
-    sdpCachedTargetTupleList.Reset();
-    sdpCachedTargetPrefixTupleList.Reset();
-    sdpCachedTargetSuffixTupleList.Reset();
-    vector<int>().swap(sdpCachedMaxFragmentChain);
-    vector<double>().swap(probMat);
-    vector<double>().swap(optPathProbMat);
-    vector<float>().swap(lnSubPValueMat);
-    vector<float>().swap(lnInsPValueMat);
-    vector<float>().swap(lnDelPValueMat);
-    vector<float>().swap(lnMatchPValueMat);
-    vector<int>().swap(clusterNumBases);
-  }
-};
 
 
 int CountZero(unsigned char *ptr, int length) {
